requirements.txt
langchain
langchain-classic
langchain-openai
langchain-google-genai
langchain-groq
langchain-community
langchain-text-splitters
langgraph
chromadb
streamlit
python-dotenv
pdf2image
pytesseract
pdfplumber
python-docx
beautifulsoup4
requests
sentence-transformers
